+++ emg3d/models.py
@@ -580,6 +580,9 @@ class VolumeModel:
         # Store case, for restriction.
         self.case = model.case
 
+        # Store a minimal TensorMesh.
+        self.grid = meshes._TensorMesh(model.grid.h, model.grid.origin)
+
         # eta_x
         self._eta_x = self.calculate_eta('property_x', model, sfield)